apps/neovim: Fix statusline issue with split windows
TL;DR: neovim really needs a version of the '%!' statusline %-expression
that runs the specified expression in the context of the window that the
statusline belongs to, rather than in the context of the
currently-active window and buffer.

Without it, we have to awkwardly hack around this lack by taking
advantage of '%{}' expressions being evaluated within the context of the
window that the statusline belongs to, but they have the limitation that
their %-expressions in their output aren't evaluated.